Test again

+++ main.c
@@ -8,7 +8,7 @@ int soNguyenTo(int soA)
     if (soA < 2)    
         return 0;
 
-    for (int i = 2; i <= sqrt((float)soA); i ++)
+    for (int i = 2; i <= sqrt(soA); i ++)
     {
         if (soA%i==0)
         {
